+++ projects/website/src/app/helper.service.ts
@@ -4,5 +4,6 @@ import type { ToasterPosition } from 'better-toast';
 @Injectable({ providedIn: 'root' })
 export class HelperService {
   readonly position = signal<ToasterPosition>('bottom-right');
+  readonly stacked = signal(true);
   readonly richColors = signal(false);
 }